pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "genai-contract-generator"
version = "0.1.0"
description = "GenAI-Powered Data Contract Generator using Claude and FastAPI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "anthropic>=0.40.0",
    "sqlglot>=23.0.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
    "structlog>=24.0.0",
    "python-dotenv>=1.0.0",
    "jsonschema>=4.22.0",
    "starlette>=0.37.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-mock>=3.14.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "respx>=0.21.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=. --cov-report=term-missing --cov-fail-under=85"

[tool.coverage.run]
omit = ["tests/*", "ui/*", "examples/*", "integrations/anthropic_client.py", "integrations/datacontract_cli.py"]

[tool.hatch.build.targets.wheel]
packages = ["api", "core", "integrations", "models"]
